+++ linking-platform/smart-card/src/common/ui/icons/live-document-icon.tsx
@@ -1,0 +1,8 @@
+import LiveDocumentIcon from '@atlaskit/icon-lab/core/page-live-doc';
+
+import { renderIconTile } from './utils';
+
+const LiveDocumentIconWithColor = renderIconTile(LiveDocumentIcon, 'magentaBold');
+LiveDocumentIconWithColor.displayName = 'LiveDocumentIconWithColor';
+
+export default LiveDocumentIconWithColor;