+++ editor/editor-plugins/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-plugins
 
+## 7.2.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 7.2.2
 
 ### Patch Changes